Rewire-ui: - Fix isEmail and isRegex validators to convert an undefined value to an empty string, so that the regex tests work for undefined values when an empty value is a valid match.     - Add a default validator to 'phone' fields, that makes sure that if the field isn't empty, it has the correct length. This length is determined by the 'format' of this phone number field.     - Check Field inner container css changes for alignment fix in rewire-grid.
Rewire-grid: - Fix for grid cell left / right navigation while editing a cell. While editing, these keys will now only traverse the value being edited, and no longer prompt to go to the previous / next cell. Shift + Tab and Tab still accomplishes this while editing. The reason for this change is because when attempting to navigate a string value to edit, an accidental extra tap of the left or right arrow key while at the beginning or the end of the string is not an uncommon occurence, and should not produce such undesirable behaviour.