Rewire-ui: New IFieldDefn function to use when creating form fields: onValueChange((v: any) => void). Gives custom callbacks to form fields when their values change, (in addition to the base one that simply sets the fields value to the input value). Allows for things like setting fields to the sum of other fields, while keeping it updated to the right value if the others change, etc. Had to do this with watchers before, which wasn't as ideal.
Rewire-grid: - Added similar onValueChange custom callback functionality to grid columns / cells. Set the onValueChange callback member of the columnModel / cellModel desired, and it will be called whenever a cell in that column / that specific cell is given a new value. Very useful for things like making a columns cells equal to the sum of a few previous cells, etc, even after those other cells have changed values on the fly.
    - Cells will now properly render falsey values that are not undefined or null.